tests/justhtml-tests/in_body_recovery.dat

#data
<dl><dd>x</dd></dl>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <dl>
|       <dd>
|         "x"

#data
<p><u><section><meta>
#errors
(1,3): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <p>
|       <u>
|     <section>
|       <meta>

#data
<marquee></body><!
#errors
(1,9): expected-doctype-but-got-start-tag
(1,10): unexpected-end-tag
(1,18): eof-in-tag

#document
| <html>
|   <head>
|   <body>
|     <marquee>
|       <!--  -->

#data
<div><tt></div><rt>
#errors
(1,5): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <div>
|       <tt>
|     <rt>

#data
<form><table></form><table>
#errors
(1,6): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <form>
|       <table>
|       <table>


#data
<p><b>x</p>y</b>
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <b>
|         "x"
|     <b>
|       "y"

#data
<rt><rt>
#errors
(1,4): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <rt>
|       <rt>

#data
<f><frameset>
#errors
(1,3): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <f>

#data
<div><b>x</div><span>y</span><svg>z</svg>
#errors
(1,5): expected-doctype-but-got-start-tag
#document
| <html>
|   <head>
|   <body>
|     <div>
|       <b>
|         "x"
|     <b>
|       <span>
|         "y"
|       <svg svg>
|         "z"

#data
<a><dialog><a>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|       <dialog>
|     <a>

#data
</r\x0a=="><
#errors
#document
| <html>
|   <head>
|   <body>

#data
<p><a><xmp>
#errors
#document
| <html>
|   <head>
|   <body>
|     <p>
|       <a>
|     <a>
|       <xmp>

#data
<!DOCTYPE html><li><menuitem><li>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <li>
|       <menuitem>
|     <li>

#data
<object><h2><button><ul></ul></h2><iframe>x
#errors
#document
| <html>
|   <head>
|   <body>
|     <object>
|       <h2>
|         <button>
|           <ul>
|       <iframe>
|         "x"
